% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{autocovmat}
\alias{autocovmat}
\title{Autocovariance matrix}
\usage{
autocovmat(kernel, x)
}
\arguments{
\item{kernel}{An object representing a gaussian process' kernel, e.g. a
\emph{sde_kernel} object.}

\item{x}{A matrix in which each row represents an input vector. Note that
if the inputs are univariate, a matrix with just one column should
be used.}
}
\description{
Computes the autocovariance matrix of the input vectors represented with
a matrix (in which each row represents an input vector).
}

