% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp_sde.R
\name{constant_function_factory}
\alias{constant_function_factory}
\title{Create a function returning a constant value}
\usage{
constant_function_factory(v)
}
\arguments{
\item{v}{Constant value to be returned by the function}
}
\value{
A function returning a vector with constant value \emph{v} and
length equal to the input vector
}
\description{
Create a function returning a constant value
}
\examples{
fun = constant_function_factory(0)
# all values are TRUE
print(fun(1:10) == rep(0, 10))
}

