% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{covmat}
\alias{covmat}
\title{Covariance matrix}
\usage{
covmat(kernel, x, y)
}
\arguments{
\item{kernel}{An object representing a gaussian process' kernel, e.g. a
\emph{sde_kernel} object.}

\item{x, y}{A matrix in which each row represents an input vector. Note that
if the inputs are univariate, a matrix with just one column should
be used.}
}
\description{
Computes the covariance matrix of the input vectors represented with
matrices (in which each row represents an input vector).
}

