% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{do_events}
\alias{do_events}
\title{Dansgaard-Oeschger (DO) events time series}
\format{A \emph{ts} object}
\source{
\url{ftp://ftp.ncdc.noaa.gov/pub/data/paleo/icecore/greenland/summit/ngrip/isotopes/ngrip-d18o-50yr.txt}
}
\usage{
do_events
}
\description{
A time series of delta-O-18 measurements (interpreted as temperature proxy)
during  the last glacial period and showing abrupt climate fluctuations
known as Dansgaard-Oeschger (DO) events.
}
\keyword{datasets}

