% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbr_sde.R
\name{fit_kbr_sde}
\alias{fit_kbr_sde}
\title{Kernel Based SDEs Estimation}
\usage{
fit_kbr_sde(x, h, kernels = c("normal", "normal"), driftBw = 0.5,
  diffBw = 0.5, nSim = 500, nthreads = 1, solveTiesDrift = c("maxDrop",
  "minArg", "maxArg", "minVal", "maxVal", "na"), solveTiesDiff = c("maxDrop",
  "minArg", "maxArg", "minVal", "maxVal", "na"), driftErrorBw = 0.1,
  diffErrorBw = driftErrorBw, plotErrors = TRUE)
}
\arguments{
\item{x}{A univariate vector representing the time series}

\item{h}{The sampling period of the time series}

\item{kernels}{A vector of 2 strings specifying which kernels should be used
in the KBR estimation for the drift (first component of the vector) and the
diffusion (second component). It is currently ignored since only the Gaussian
kernel is supported.}

\item{driftBw, diffBw}{Bandwidth of the Gaussian kernel used to estimate the
drift/diffusion term}

\item{nSim}{= 500 Number of simulations used to calculate the delta-error
(see references)}

\item{nthreads}{= 1 Number of threads to be used during the computation}

\item{solveTiesDrift, solveTiesDiff}{A string specifying the strategy to
be used to break ties between two different bandwidths with the same
delta-error for the drift and diffusion terms, respectively.}

\item{driftErrorBw, diffErrorBw}{Since the delta-errors are noisy, a rolling
mean is used before selecting the best bandwidth. These parameters specify
the width of the rolling mean (in units of bandwidth) used to smooth the
drift-errors and the diffusion-errors, respectively}

\item{plotErrors}{Boolean value. Plot delta-errors?}
}
\description{
Estimates the drift and diffusion terms of a Langevin Equation using the
Kernel Based Regression (KBR) method from a one-dimensional time series.
}

