% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_sde.R
\name{fit_polynomial_sde}
\alias{fit_polynomial_sde}
\title{Polynomial based SDEs Estimation}
\usage{
fit_polynomial_sde(timeSeries, samplingPeriod = NULL, timeTs = NULL,
  nDrift = 1, nDiff = 1, direction = "both")
}
\arguments{
\item{timeSeries}{A univariate vector representing the time series}

\item{samplingPeriod}{The sampling period of the time series}

\item{timeTs}{A vector with the occurrence time of each sample of the time
series. Can be used as an alternative to \emph{samplingPeriod}.}

\item{nDrift, nDiff}{Maximum order of the polynomial used to fit the
drift/diffusion term.}

\item{direction}{The mode of stepwise search, can be one of "both",
"backward", or "forward". See \code{\link[stats]{step}}.}
}
\value{
A list with two \emph{poly_sde} objects representing the estimates
for the drift and the diffusion terms. The \emph{poly_sde} objects can be
used with \emph{predict} to get numerical estimates of the drift/diffusion
functions.
}
\description{
Polynomial based SDEs Estimation
}

