% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{get_hyperparams}
\alias{get_hyperparams}
\title{Get the hyperparameters of a kernel}
\usage{
get_hyperparams(kernel)
}
\arguments{
\item{kernel}{An object representing a gaussian process' kernel, e.g. a
\emph{sde_kernel} object.}
}
\description{
Get the hyperparameters of a kernel
}

