% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gp_kernels}
\alias{gp_kernels}
\title{Gaussian process' kernels}
\description{
Available kernels for creating gaussian processes
}
\details{
\itemize{
\item \emph{gp_kernel} Abstract class representing a generic GP kernel. Final users
should not use it.
\item \emph{exp_kernel} An squared-exponential kernel with fixed amplitude.
\item \emph{rq_kernel} A rational-quadratic kernel with fixed amplitude.
\item \emph{exp_const_kernel} A sum of a squared-exponential kernel and a constant
value. The sum of the constant value and the amplitude of the exponential
kernel are tied so that they always sum the same value.
\item \emph{sum_exp_kernels} A sum of two different squared-exponential kernels:
The sum of the amplitudes of the exponential
kernels are tied so that they always sum the same value.
\item \emph{clamped_exp_lin_kernel} A sum of a squared-exponential kernel and a
linear kernel. The sum of the amplitudes of the two terms are tied so
that they always sum the same value. Note that this kernel is non-stationary.
}
}

