% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{ornstein}
\alias{ornstein}
\title{Ornstein-Uhlenbeck time series}
\format{A \emph{ts} object}
\usage{
ornstein
}
\description{
A simulation of a trajectory from an Ornstein-Uhlenbeck SDE with drift
f(x) = -x and diffusion sqrt(g(x)) = sqrt(1.5)
}
\keyword{datasets}

