% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{predict.voila_preproc}
\alias{preproc}
\title{Preprocess a time series}
\usage{
preproc(x, method = c("center", "scale"), useCenter = NULL,
  useScale = NULL)

\method{predict}{voila_preproc}(object, input, inverse = FALSE, ...)
}
\arguments{
\item{x}{time series to preprocess}

\item{method}{Vector with the preprocessing methods to apply.}

\item{useCenter, useScale}{Numeric value specifying the center/scale to use
for the preprocessing instead of learning these values from the time
series \emph{x}}

\item{object}{A \emph{voila_preproc} object}

\item{input}{The input to be transformed using \emph{object}}

\item{inverse}{Logical value: Is the data being trasformed back?}

\item{...}{Additional parameters (currently ignored)}
}
\value{
\emph{preproc} returns a  \emph{voila_preproc} object that can
be used to transform new data using \emph{predict}.
}
\description{
Center and scale a time series
}
\examples{
data("do_events")
plot(do_events, main = "DO events")
pp = preproc(do_events)
scaled_do_events = predict(pp, do_events)
plot(scaled_do_events, main = "Scaled DO events")
# recover the original DO events
original_do_events = predict(pp, scaled_do_events, inverse = TRUE)
}

