% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{sde_kernel}
\alias{sde_kernel}
\title{Create a gaussian process' kernel}
\usage{
sde_kernel(type = c("exp_kernel", "rq_kernel", "sum_exp_kernels",
  "exp_const_kernel", "clamped_exp_lin_kernel"), parameters,
  inputDimension = 1, epsilon = 0)
}
\arguments{
\item{type}{A string specifying the type of kernel to create}

\item{parameters}{A NAMED list with the parameters that the kernel needs for
its proper creation. The easiest way of seeing which parameters are required
is to make this function fail. For example:
\emph{sde_kernel("exp_kernel",list())}}

\item{inputDimension}{The input dimension of the kernel}

\item{epsilon}{A small value to be added to the diagonal of the kernel
covariance matrices to regularize them. This improves the numerical stability
of the computations.}
}
\value{
A \emph{sde_kernel} S3 object
}
\description{
Creates a kernel defining the properties of a gaussian process
}
\examples{
# See demo/ornstein for a complete example
data("ornstein")
uncertainty = 5
inputDim = 1
# A small value to regularize covariance matrices
epsilon = 1e-5
# Create a exponential kernel with lengthScale = 1 and amplitude = uncertainty
# to model the drift function. We may select the amplitude from the fact that it
# is a gaussian process and the 95\% confidence interval would be
# (-2 * sqrt(amplitude), 2 * sqrt(amplitude)).
driftKer = sde_kernel("exp_kernel",
                      list('amplitude' = uncertainty,
                           'lengthScales' = 1),
                      inputDim, epsilon)
# The selection of the amplitude parameters for gaussian process modelling the
# diffusion term is more complicated since voila actually uses a lognormal
# process. This function helps with the selection of the parameters from
# an uncertainty parameter and the time series
diffPars = select_diffusion_parameters(ornstein, deltat(ornstein),
                                       priorOnSd = uncertainty)
# Create another exponential kernel with lengtScale=1
diffKer = sde_kernel("exp_kernel",
                      list('amplitude' = diffPars$kernelAmplitude,
                           'lengthScales' = 1),
                      inputDim, epsilon)
}
\seealso{
\code{\link{covmat}}, \code{\link{autocovmat}}, \code{\link{vars}},
\code{\link{get_hyperparams}}, \code{\link{set_hyperparams}},
\code{\link{decrease_upper_bound}} and \code{\link{increase_lower_bound}}
}

