% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_diffusion_parameters.R
\name{select_diffusion_parameters}
\alias{select_diffusion_parameters}
\title{Parameter selection for the diffusion term}
\usage{
select_diffusion_parameters(x, samplingPeriod, priorOnSd,
  responseVariableIndex = 1, varX = NULL)
}
\arguments{
\item{x}{A time series that may be modelled with a Langevin equation.}

\item{samplingPeriod}{the sampling period of the time series}

\item{priorOnSd}{An estimate for the maximum change in amplitude of the
diffusion term.}

\item{responseVariableIndex}{Integer indicating in which of the input
dimensions the inference will be made.}

\item{varX}{The selection of the hyperparametes is based on the variance
of the differentiated time series. Although it is estimated by the procedure,
the user may specify a different value using this parameter.}
}
\value{
A list with a mean (\emph{v}) and a covariance amplitude for the
log-gaussian process (\emph{kernelAmplitude}).
}
\description{
The selection of the amplitude parameters for gaussian process modelling the
diffusion term is complicated since we have to link the covariance matrix of
a log-normal random variable with our prior belief about the \emph{variance}
of the diffusion term (here, variance may be interpreted as a reasonable
estimate for the squared-amplitude of the diffusion). This function helps
with the selection of those parameters.
}

