% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{set_hyperparams}
\alias{set_hyperparams}
\title{Set the hyperparameters of a kernel}
\usage{
set_hyperparams(kernel, hyperparams)
}
\arguments{
\item{kernel}{An object representing a gaussian process' kernel, e.g. a
\emph{sde_kernel} object.}

\item{hyperparams}{The new hyperparameters of the kernel}
}
\description{
Set the hyperparameters of a kernel
}

