% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp_sde.R
\name{sgp_sde}
\alias{sgp_sde}
\title{SGP-based estimate of a Langevin Equation/SDE}
\usage{
sgp_sde(kernel, xm, priorMeanFun, posteriorMean, posteriorCov)
}
\arguments{
\item{kernel}{@param kernel An object representing a gaussian process'
kernel, e.g. a \emph{sde_kernel} object.}

\item{xm}{A matrix representing the pseudo-inputs of the Sparse Gaussian
Process (SGP).}

\item{priorMeanFun}{A function returning the prior mean of the gaussian
process.}

\item{posteriorMean}{A vector representing the mean of the gaussian
distribution that approximates the posterior of the drift/diffusion term.}

\item{posteriorCov}{A matrix representing the covariance matrix of the
gaussian distribution that approximates the posterior of the
drift/diffusion term.}
}
\description{
SGP-based estimate of a Langevin Equation/SDE
}
\seealso{
\code{\link{predict.sgp_sde}}
}

