% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sde.R
\name{simulate_sde}
\alias{simulate_sde}
\title{Simulate Langevin Equations}
\usage{
simulate_sde(driftExpression, diffExpression, samplingPeriod, tsLength, xinit,
  trueParameter = list())
}
\arguments{
\item{driftExpression, diffExpression}{A string/\emph{expression} specifying
the drift and diffusion terms}

\item{samplingPeriod}{The sampling period of the resulting time series}

\item{tsLength}{The length of the resulting time series}

\item{xinit}{Initial value for the resulting time series}

\item{trueParameter}{A list containing the true parameters of the drift and
diffusion expressions (in the case that there are some unknown values).}
}
\value{
A matrix containing the simulated trajectory.
}
\description{
A simple wrapper for yuima's SDE simulation methods
}
\examples{
simTs = simulate_sde("-x", "sqrt(2)", 0.001, 1000)
plot.ts(simTs, xlab = "Time t", ylab = "x(t)",
        main = "Ornstein-Uhlenbeck process")
}
\seealso{
\code{\link[yuima]{simulate}}
}

